Adapt tests to new structure.
In previous structure, NDN lite tests were in their own repository
with NDN lite sources as a submodule in base directory. Here, the
relationship is inverse, i.e., the tests are integrated directly in
NDN lite repository.

In this commit a really cheap trick (symlink to parent instead of
submodule) is used to have the tests running without changing the
source code. This might need some reconsideration in the future!